change locations of deduplication functions
