+++ src/lib/db.ts
@@ -1,21 +1,13 @@
 import { PrismaClient } from '@prisma/client';
-import { createClient } from 'redis';
 
-// add prisma and redis to the NodeJS global type
+// add prisma to the NodeJS global type
 interface CustomNodeJsGlobal extends NodeJS.Global {
   prisma: PrismaClient;
-  redis: ReturnType<typeof createClient>;
 }
 
-// Prevent multiple instances of databases in development
+// Prevent multiple instances of Prisma Client in development
 declare const global: CustomNodeJsGlobal;
 
 export const db = global.prisma || new PrismaClient();
-export const redis = global.redis || createClient();
 
-redis.connect();
-
-if (process.env.NODE_ENV === 'development') {
-  global.prisma = db;
-  global.redis = redis;
-}
+if (process.env.NODE_ENV === 'development') global.prisma = db;
