+++ src/lib/index.ts
@@ -1,5 +1,6 @@
 export * from './debate-tools';
 export * from './db';
+export * from './redis';
 export * from './events';
 
 // XXX: hacky solution for typesafe async pipe
